+++ llm_ffw/__init__.py
@@ -6,6 +6,7 @@ from .capabilities import (
     EmailAddressCapability,
     JSONOutputCapability,
     IPAddressCapability,
+    MACAddressCapability,
     RuleCapability,
     SecretCatalogCapability,
     UnsafeURLCapability,
@@ -35,6 +36,7 @@ from .inspection import (
 )
 from .json_output import JSONOutputConfig
 from .ip_address import IPAddressConfig
+from .mac_address import MACAddressConfig
 from .email_address import EmailAddressConfig
 from .unsafe_url import UnsafeURLConfig
 from .payment_card import PaymentCardConfig
@@ -81,6 +83,7 @@ from .rules.unicode_tag_smuggling import UnicodeTagSmugglingRule
 from .rules.banned_substrings import BannedSubstringsRule
 from .rules.json_output import JSONOutputRule
 from .rules.ip_address import IPAddressRule
+from .rules.mac_address import MACAddressRule
 from .rules.email_address import EmailAddressRule
 from .rules.unsafe_url import UnsafeURLRule
 from .rules.payment_card import PaymentCardRule
@@ -121,6 +124,9 @@ __all__ = [
     "IPAddressCapability",
     "IPAddressConfig",
     "IPAddressRule",
+    "MACAddressCapability",
+    "MACAddressConfig",
+    "MACAddressRule",
     "LLMFirewall",
     "LLMFirewallManager",
     "LiteralMatchMode",