+++ llm_ffw/rules/__init__.py
@@ -6,6 +6,7 @@ from .invisible_characters import InvisibleCharactersRule
 from .unicode_tag_smuggling import UnicodeTagSmugglingRule
 from .json_output import JSONOutputRule
 from .ip_address import IPAddressRule
+from .mac_address import MACAddressRule
 from .email_address import EmailAddressRule
 from .payment_card import PaymentCardRule
 from .private_key import PrivateKeyRule
@@ -19,6 +20,7 @@ __all__ = [
     "UnicodeTagSmugglingRule",
     "JSONOutputRule",
     "IPAddressRule",
+    "MACAddressRule",
     "EmailAddressRule",
     "PaymentCardRule",
     "PrivateKeyRule",